Added a ShrinkSpell / ShrinkOnTrigger.cs
@itsschwer
@primus852
Would u suggest it in this way?
I think this would be usefull but i got some problems with it.

1.) I want more things to happen. But that differs on the PlayerMovement/Enemy.
So I need overrides. Somehow I'm not able to override the IENum which would be needed...
a) If its the Player I want him to deal less damage which is calculated on the weapon.
b.) If its an Enemys I also want him to deal less damage which is depending on several scripts.

Maybe there's a smarter way :)

Have a great time! :+1: